pyproject.toml
[project]
name = "coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "baidusearch>=1.0.3",
    "boto3>=1.37.19",
    "bs4>=0.0.2",
    "duckduckgo-search>=7.5.3",
    "googlesearch-python>=1.3.0",
    "loguru>=0.7.3",
    "openai>=1.68.2",
    "pydantic>=2.10.6",
    "requests>=2.32.3",
    "tenacity>=9.0.0",
    "tiktoken>=0.9.0",
]
